+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 0.7
+* Added `showbConstPrec` (and corresponding `Show` and `Show1` instances for `Const`) to `Text.Show.Text.Control.Applicative`
 * Added `showbUArrayPrec` (and corresponding `Show` instance for `UArray`s) and `showbIArrayPrec` to `Text.Data.Text.Data.Array`.
 * Renamed `showbListDefault` to `showbListWith` to match how `Text.Show` names it
 * Exposed `showbShortByteString` with all versions of `bytestring` by using the `bytestring-builder` package
@@ -6,6 +7,7 @@
 * The Template Haskell deriver now handles showable unlifted types (`Char#`, `Double#`, `Float#`, `Int#`, and `Word#`) correctly
 * Removed `LitChar` and `LitString` from `Text.Show.Text.Data.Char`, as they were not as useful as I had imagined.
 * Removed the deprecated `replicateB` function
+* `Typable` instances for `Show`, `Show1`, and `GShow` (with GHC 7.8 and later)
 * `Typeable` instance for `ConType`
 * Only derive `Eq` and `Ord` for `ConType` if a recent-enough version of `text` is used
 * Changed the implementations of some functions in `Text.Show.Text.Debug.Trace` to use `ByteString`s instead of `String`s